upgrade Lucene and dependencies to 7.2.x

+++ rhapsode-common/src/main/java/org/rhapsode/app/handlers/search/BasicSearchHandler.java
@@ -218,7 +218,8 @@ public class BasicSearchHandler extends AbstractSearchHandler {
 
     }
 
-    private static void writeNextPrev(BasicSearchRequest basicSearchRequest, int totalHits, RhapsodeXHTMLHandler xhtml) throws SAXException {
+    private static void writeNextPrev(BasicSearchRequest basicSearchRequest,
+                                      long totalHits, RhapsodeXHTMLHandler xhtml) throws SAXException {
         if (totalHits < 2) {
             return;
         }
